Relocate test-only blocking Producer client to test support
